refactor: move routes constants to separate file
