+++ src/components/ProfileNav/ProfileProjectsNav/ProfileProjectsNav.tsx
@@ -1,7 +1,7 @@
 import cl from 'classnames';
 import { FC } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import { ROUTE_PROFILE_PROJECTS } from '../../../constants/constants';
+import { ROUTE_PROFILE_PROJECTS } from '../../../constants/routesConstants';
 import { TProjectShortInfo } from '../../../types/data';
 import { ArrowFatRightIcon } from '../../ui/icons';
 import { ProfileNavLink } from '../ProfileNavLink/ProfileNavLink';
