+++ src/pages/CompetenciesPage/CompetenciesPage.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { Outlet, useMatch } from 'react-router-dom';
 import { Main } from '../../components/Main/Main';
-import { ROUTE_COMPETENCIES } from '../../constants/constants';
+import { ROUTE_COMPETENCIES } from '../../constants/routesConstants';
 import { BottomGridSection } from './BottomGridSection/BottomGridSection';
 import { TopGridSection } from './TopGridSection/TopGridSection';
 import s from './CompetenciesPage.module.scss';
